requirements.txt
# --------- pytorch --------- #
torch==2.9.0
torchvision==0.24.0
lightning==2.5.5
torchmetrics==1.8.2
tiktoken==0.12.0
transformers==4.57.1
datasets==4.2.0
modelscope==1.31.0
gdown==5.2.0
sacrebleu==2.5.1
sentencepiece==0.2.1

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
tensorboard


# --------- others --------- #
rootutils       # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
gpustat
